+++ src/components/common/bookList/style.ts
@@ -1,9 +1,4 @@
-import { colorStyle } from "@/styles/colorStyle";
 import styled from "@emotion/native";
-
-interface BarProps {
-  width: number
-}
 
 export const Container = styled.View`
   flex: 1;
@@ -19,10 +14,6 @@ export const BookBox = styled.View`
   gap: 16px;
 `
 
-export const BookBar = styled.View<BarProps>`
-  background-color: ${colorStyle.defaultWhite};
-  width: ${({ width }) => width}px;
-  height: 16px;
-  border-bottom-left-radius: 8px;
-  border-bottom-right-radius: 8px;
+export const NotFound = styled.View`
+  height: 173px;
 `
